% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_values.R
\name{build_starting_values}
\alias{build_starting_values}
\title{Build starting values for the sampler}
\usage{
build_starting_values(N, K, lag, fixed_list = NULL)
}
\arguments{
\item{N}{Integer. Number of  variables.}

\item{K}{Integer. State dimension of the companion form (K = N * lag)
of the state matrix.}

\item{lag}{Integer. Maximum autoregressive lag of the VAR.}

\item{fixed_list}{Optional list. If provided, it can contain any
of the elements \code{m}, \code{sigma2m}, \code{sw}, \code{aw},
\code{ss}, \code{nus}, \code{as}, \code{bs}. Each non-NULL entry
replaces the corresponding default value in the output.}
}
\value{
A list containing starting values for:
\itemize{
\item \code{M}  – K × N mean matrix for global parameter A.
\item \code{S}, \code{SS} – N × N covariance-related matrices.
\item \code{W} – K × K diagonal matrix based on decreasing weights based on lags.
\item fixed degrees of freedom: \code{mu_Sigma}, \code{eta}, \code{nu} chosen to ensure finite
first moment for inverse-Wishart and Wishart distributions.
And the elements possibly taken from \code{fixed_list}.
}
}
\description{
Creates a list of initial parameter values and hyperparameters
needed to start the sampler. The function builds
defaults for matrices \code{M}, \code{S}, \code{SS}, \code{W},
and scalar hyperparameters, optionally replacing some
of them with user-provided fixed values.
}
\examples{
init <- build_starting_values(N = 5, K = 10, lag = 2)
}
