% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_compute.R
\name{compute_global_irf}
\alias{compute_global_irf}
\title{Compute global structural IRFs (VAR(1))}
\usage{
compute_global_irf(chains, H)
}
\arguments{
\item{chains}{A list of MCMC chains from \code{run_mcmc()$CHAINS}.}

\item{H}{Number of horizons.}
}
\value{
A 4D array:
\code{Phi_global[draw, horizon, response, shock]}.
}
\description{
Computes global-level structural impulse responses from posterior
draws of the global VAR parameters (\code{A} and \code{Sigma}).
}
