% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{invwishart_rnd}
\alias{invwishart_rnd}
\title{Inverse-Wishart generator}
\usage{
invwishart_rnd(df, S)
}
\arguments{
\item{df}{Integer. Degrees of freedom (df ≥ dimension of S).}

\item{S}{Matrix. Scale matrix (K × K), symmetric and positive-definite.}
}
\value{
A numeric positive-definite matrix drawn from IW(df, S).
}
\description{
Generates a draw from an inverse-Wishart distribution with
degrees of freedom \code{df} and scale matrix \code{S}.
}
\examples{
S <- diag(3)
invwishart_rnd(10, S)

}
