% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rmatnorm}
\alias{rmatnorm}
\title{Matrix-normal generator}
\usage{
rmatnorm(M, V, S)
}
\arguments{
\item{M}{Matrix. Mean matrix of size K × N.}

\item{V}{Matrix. Row covariance matrix (K × K), positive-definite.}

\item{S}{Matrix. Column covariance matrix (N × N), positive-definite.}
}
\value{
A numeric matrix of dimension K × N.
}
\description{
Generates a draw from a matrix-normal distribution
with mean matrix \code{M}, row covariance \code{V}, and
column covariance \code{S}.
}
\examples{
M  <- matrix(0, 3, 2)
V  <- diag(3)
S  <- diag(2)
rmatnorm(M, V, S)

}
