% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{wishart_rnd}
\alias{wishart_rnd}
\title{Wishart generator}
\usage{
wishart_rnd(df, S)
}
\arguments{
\item{df}{Integer. Degrees of freedom (df ≥ dimension of S).}

\item{S}{Matrix. Scale matrix (K × K), symmetric and positive-definite.}
}
\value{
A numeric positive-definite matrix drawn from W(df, S).
}
\description{
Generates a random draw from a Wishart distribution with
degrees of freedom \code{df} and scale matrix \code{S}.
}
\examples{
S <- diag(3)
wishart_rnd(10, S)

}
